Put fee test in FeeUtilsTest.java
